+++ Src/MoneyFox.Ui.Tests/ViewModels/Budget/EditBudgetViewModelTests.cs
@@ -13,7 +13,7 @@ using MediatR;
 using Messages;
 using NSubstitute;
 using Views.Budget;
-using Views.Budget.ModifyBudget;
+using Views.Budget.BudgetModification;
 using Views.Categories;
 using Views.Categories.CategorySelection;
 using Xunit;
