+++ Src/MoneyFox.Ui/Views/Budget/BudgetModification/AddBudgetPage.xaml
@@ -4,13 +4,13 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:resources="clr-namespace:MoneyFox.Ui.Resources.Strings"
-    xmlns:modifyBudget="clr-namespace:MoneyFox.Ui.Views.Budget.ModifyBudget"
+    xmlns:budgetModification="clr-namespace:MoneyFox.Ui.Views.Budget.BudgetModification"
     Title="{x:Static resources:Translations.AddBudgetTitle}"
-    x:DataType="modifyBudget:AddBudgetViewModel"
-    x:Class="MoneyFox.Ui.Views.Budget.ModifyBudget.AddBudgetPage">
+    x:DataType="budgetModification:AddBudgetViewModel"
+    x:Class="MoneyFox.Ui.Views.Budget.BudgetModification.AddBudgetPage">
 
     <Grid RowDefinitions="*, 6, 40" Padding="12">
-        <modifyBudget:ModifyBudgetView Grid.Row="0" />
+        <budgetModification:ModifyBudgetView Grid.Row="0" />
         <Button Grid.Row="2"
                 Text="{x:Static resources:Translations.SaveLabel}"
                 Command="{Binding SaveBudgetCommand}"
