+++ Src/MoneyFox.Ui/Views/Budget/BudgetModification/EditBudgetPage.xaml
@@ -4,14 +4,14 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:resources="clr-namespace:MoneyFox.Ui.Resources.Strings"
-    xmlns:modifyBudget="clr-namespace:MoneyFox.Ui.Views.Budget.ModifyBudget"
+    xmlns:budgetModification="clr-namespace:MoneyFox.Ui.Views.Budget.BudgetModification"
     Title="{x:Static resources:Translations.EditBudgetTitle}"
-    x:DataType="modifyBudget:EditBudgetViewModel"
-    x:Class="MoneyFox.Ui.Views.Budget.ModifyBudget.EditBudgetPage"
+    x:DataType="budgetModification:EditBudgetViewModel"
+    x:Class="MoneyFox.Ui.Views.Budget.BudgetModification.EditBudgetPage"
     Visual="Material">
 
     <Grid RowDefinitions="*, 6, 40, 6, 40" Padding="12">
-        <modifyBudget:ModifyBudgetView />
+        <budgetModification:ModifyBudgetView />
         <Button Grid.Row="2"
                 Text="{x:Static resources:Translations.SaveLabel}"
                 Command="{Binding SaveBudgetCommand}"
